+++ tests/resources/test_analysis.py
@@ -16,7 +16,7 @@ from tests.base import BaseTestCase
 class AnalysisTestCase(BaseTestCase):
     def test_instantiate_analysis(self):
         """Ensures that the base analysis class can be instantiated"""
-        # from octue import runner  # <-- instantiated in the library, not here
+        # from octue.twined import runner  # <-- instantiated in the library, not here
         analysis = Analysis(twine="{}")
         self.assertEqual(analysis.__class__.__name__, "Analysis")
